day8/autoquant-alpha-day8/requirements.txt
alpaca-py>=0.20.0
numpy>=1.26.0
rich>=13.0.0
pytest>=7.4.0
python-dotenv>=1.0.0
